Made a public dataplot fuction
This plots the data and simulation for a particular file and compound. This
used to be a private function, but decided to make it public for interactive
use.

Very important change. My gcms data extractor adds the simplified file name to
the returned gcms object, which should make things much easier. In addition,
it does not require the full path to the name, because path info is stripped
before searching. This saves a lot of typing.