Fix integration
The integration routine would not work for an internal standard. I made a
large number of changes to fix this problem; however, things are more logical
now.

Big changes. Modified the _integrate method of the Nnls fitter to calculate
the integrals for reference compounds that have a START and STOP flag in their
metadata. This meant that these flags needed to be added to ref_specs.txt.
This meant that the _int_extract method of the AiaFile could also be
simplified to return a single integral for every reference compound. I needed
to change the calibration routines as well to handle the changes to retreaving
the integrals. This has the benefit that integral determination is now part of
the fitting routine rather than the calibration routine, which makes more
logical sense, and removes some custom logic from the Calibrate class.

Added a bunch of comments to the code. Updated calibration.csv because the
start and stop values are not necessary any longer.