Deepcopy ref_meta dictionary
I had a major/subtle error in the referencing object. I was copying the
ref_meta dictionary, which wasn't making copies of its underlying
dictionaries. That meant that when I was fitting bunches of data, the
integrals were always equal to the last compound in the bunch.

Imported the copy module from the standard library, and used the deepcopy
function instead, which copies all the internal dictionaries as well. This
fixes the problem.