Folder check and start/stop time fix
Several of the plotting functions were not looking in the correct place for
the start/stop times and there was not check that the requested output folder
exists, which would have thrown an error. In addition, the "fitsplot" method
was not being used when plots were requested, and the call signature was not
equivalent to the other plotting methods.