Made GcmsStore a subclass of pandas.HDFStore
I don't know why I didn't do this in the first place. It required a lot of
code changes, but in the end, it actually makes things much simpler. Now I
don't have to refer to 'pdh5' and 'tbh5' all the time. Unfortunately, if I
need to use some PyTables functions, I need to refer to the _handle attribute.

Now all of the documenation needs to be updated.