Update Isotope class for changes to GcmsStore
The recent changes made to GcmsStore required some reworking of the Isotope
calls to that class.

In addition, I noticed that the tables I was creating were not compressed. I
added "filters" to all create_group calls, which ensures that all children are
compressed with the _filters defined by GcmsStore.